# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "MX 사업부 위키 백엔드. DocumentJSON v1.0 (`packages/shared/schemas/document.json`) 을 단일 진실 공급원(SSOT) 으로 사용한다.\n\n**핵심 흐름**: `POST /api/v1/documents` 한 번으로 본문 검증 → 섹션 번호 부여 → 위키링크/태그/용어집/감사로그/검색인덱스가 모두 자동 갱신된다.\n\n**Conditional write**: 본문 수정(PUT/PATCH) 은 `If-Match` 헤더 필수 (낙관적 동시성). GET 응답의 `ETag` 를 그대로 사용하면 된다.",
    "title": "MX White Paper API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{slug}/export.html": {
      "get": {
        "description": "DocumentJSON 본문을 자체 완비형(self-contained) HTML 로 렌더한다.\n\n- `style=namuwiki` (현재 유일 — 추후 academic/minimal 등 추가 예정)\n- `inline_images=1` 이미지 base64 인라인 (파일 크기 폭증 주의)\n- `katex=cdn` math 블록을 KaTeX CDN 으로 렌더\n- `mermaid=cdn` flow(mermaid) 블록을 mermaid CDN 으로 렌더",
        "operationId": "export_document_html_api_v1_documents__slug__export_html_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "style",
            "required": false,
            "schema": {
              "default": "namuwiki",
              "title": "Style",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inline_images",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1,
              "minimum": 0,
              "title": "Inline Images",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "katex",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Katex"
            }
          },
          {
            "in": "query",
            "name": "mermaid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mermaid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서를 단일 HTML 파일로 내보내기 (나무위키 스타일)",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/exports/docx": {
      "post": {
        "description": "DocumentJSON 본문을 Word 문서(.docx) 로 변환한다. cycle 1 의 .docx import 와 round-trip 가능 (스타일/구조를 importer 가 인식하는 형태로 출력).\n\n- Heading 1/2/3 = section level, Heading 4/5/6 = `heading-4` block.\n- markdown-lite (bold/italic/strike/underline/code/link/wiki/footnote) 보존.\n- table / list / quote / callout / code 는 네이티브 도형/스타일로.\n- chart / gantt / flow / org-chart 는 텍스트 요약 + 데이터 표 fallback.\n- speaker note (`meta.note=\"speaker:…\"`) 는 DOCX 노트 페인이 없어 drop.",
        "operationId": "export_docx_api_v1_exports_docx_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서를 Word (.docx) 로 내보내기",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/markdown": {
      "post": {
        "description": "DocumentJSON 본문을 GitHub-flavoured Markdown 으로 변환해 다운로드.\n차트/Gantt/Flow 는 mermaid 코드블록으로, 그 외 동적 블록은 텍스트 요약으로 보존된다.",
        "operationId": "export_markdown_api_v1_exports_markdown_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서를 Markdown 으로 내보내기",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/pdf": {
      "post": {
        "description": "WeasyPrint 가 설치된 환경에서만 동작한다.\n설치되지 않은 경우 501 을 돌려주며, FE 는 print-friendly 페이지로 폴백한다.",
        "operationId": "export_pdf_api_v1_exports_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서를 PDF 로 내보내기 (WeasyPrint)",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/pptx": {
      "post": {
        "description": "DocumentJSON 본문을 발표용 .pptx 로 변환한다.\n\n- level-1 섹션 1개 = 슬라이드 1장 (제목 슬라이드 포함).\n- level-2 sub-section 은 분량이 ≤500자면 inline bullets, 초과하면 별도 슬라이드로 분리.\n- chart 는 line/bar/pie 만 네이티브 차트로, 그 외는 텍스트 요약 fallback.\n- ``meta.note: \"speaker: …\"`` prefix 를 만난 블록은 슬라이드 노트에 누적.",
        "operationId": "export_pptx_api_v1_exports_pptx_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서를 PowerPoint (.pptx) 로 내보내기",
        "tags": [
          "exports"
        ]
      }
    },
...
  },
...
}